bug-2037435: Rename gcs directory to gcp.
